+++ Opera/APP/js/whatsNew.js
@@ -15,17 +15,18 @@
 */
 
 changes = [
-    "-1.0 Fresh look, added notifications (works only on 25+), fixed 'Following List' page, various improvements",
-    "-0.9.0.4 Bug fixing",
-    "-0.9.0.3 Code optimization",
-    "-0.9.0.2 Fixed sound",
-    "-0.9.0.1 Bugfix",
-    "-0.9.0.0 First publish in Opera Web Store"
+    "-1.1 changed font to PermianSans Typeface, fixed ON/OFF statements, added notify on offline, optimized network code",
+    "-1.0 fresh look, added notifications (works only on 25+), fixed 'Following List' page, various improvements",
+    "-0.9.0.4 bug fixing",
+    "-0.9.0.3 code optimization",
+    "-0.9.0.2 fixed sound",
+    "-0.9.0.1 gugfix",
+    "-0.9.0.0 first publish in Opera Web Store"
 ];
 messages = {
-    "v.0.9.0.3": {
-        "msg": "Code optimization",
-        "contain": ""
+    "v.1.1": {
+      "msg": "Various improvements",
+        "contain": "-changed font to PermianSans Typeface<br>-fixed ON/OFF statements<br>-added notify on offline<br>-optimized network code"
     }
 };